Changed initialFocusedIndex on drop downs to scroll into view
If you have a select list with more than 9 items, they won't
all fit in view.  If you have the 10th or later item selected,
upon open that item should not only be selected, but visible
too.  To accomidate this, I updated the code so that
initialFocusedIndex not only focuses the item, but scrolls
it into view.  This brought to light a few bugs that were
fixed along the way.

Issue: https://khanacademy.atlassian.net/browse/WB-1091
https://khanacademy.atlassian.net/browse/CLASS-7007

Test plan:
Verify that all unit tests work correctly.
In the Wonder Blocks documentation,
"Single select with search filter and custom styles"
example:
  * Open the drop down
  * Select an item that is initially out of view
  * Re-open drop down
  * Verify that the item visible
Change the example so no filter is available and
rerun the changes.

Open the dropdown
Select the filter
Type values
Enter a space, verifying that it works
Press tab, verifying that it moves you to the X