Get CJS output so that it's best for test use
We only use the CommonJS output to support Jest testing.
This update has the babel output for webpack set to target node instead
of esmodules, and also sets webpack to avoid module concatenation. This
second change is necessary to prevent webpack from mangling react
component names (it does this to avoid collisions across the concatenated
modules).