Update action-menu to pass down the opened state of the menu to the opener
Summary:
When the opener isn't updated to be opened it causes the aria-expanded prop to remain false, which hints to screen readers that the menu is not opened. This fixes that by passing the value of the state instead of the value of the opened prop.

Issue: LP-9477

Test plan:
yarn jest