Improve navigation callback documentation
This commit adds additional documentation describing the `onClick`,
`beforeNav`, and `safeWithNav` props. These props have been available in
Wonder Blocks for a while, but there has been some confusion about when to
use each one. This commit attempts to add enough detail about these
props to make it clear what the use case is for each prop.

Additionally, this commit changes the section headings in button.md from
h4 to h3 elements. This avoids a gap in heading levels and also brings the
section in line with other sections on the page.

Issue: none

Test Plan:
- Load the documentation and read the Button, Clickable, and Link
  sections.