Allow disabled buttons to be focusable by screenreaders
"There might be controls like a button that's disabled and we want to be able to communicate why it's disabled with a tooltip.  We want that tooltip to be accessible even if the control is disabled.  Disabled controls may not tabbable but users can still navigate to them via Voice Over and other screen readers."

- Change `disabled` prop to `aria-disabled`
- Change tabIndex to 0
- We don't have to add an eventListener to stop clicks from happening because that is already handled here:
  https://github.com/Khan/wonder-blocks/blob/master/packages/wonder-blocks-clickable/src/components/clickable-behavior.js#L213