Implement useInternalData
This is a corrollary of the `InternalData` component and will.
This also modifies the `Result` type to use a `status` value so that
the data is validated as loading, success, or error up front.

That makes this a breaking change, but that should be easy to cope with
given that WB Data is as yet not widely adopted.