Remove client-side caching architecture and fix hook
Good job I decided to remove the client-side caching option as it turned out that by not having it, we were secretly introducing a bug in the hook. This addresses all that and gets things working in a manner that makes sense. Any uses of the removed features can be migrated to an alternative when we update webapp.