Implement useInternalData (#1280)
## Summary:
This hook does all the work that the `InternalData` component did.

This also modifies the `Result` type to use a `status` value so that
the data is validated as loading, success, or error up front.

Finally, this removes the client-side caching.

Those last two things make this a breaking change, but that should be easy to cope with
given that WB Data is as yet not widely adopted.

I wanted to try and split this into smaller PRs but it was not so easy to do given the order in which I had been doing the work. However, since much of the change is deleting code and the only new code is either tests or the `useDataInternal` hook, I felt that it wouldn't complicate things terribly.

### Next up
The PR after this one will then implement the `useData` hook, supporting
the cache and request interception context as per the current `Data`
component and we can then replace the `Data` component with one that
uses the hooks. The `InternalData` component will go away in the next PR.

Issue: FEI-4219

## Test plan:
`yarn test`

Author: somewhatabstract

Reviewers: somewhatabstract, kevinbarabash, jeresig, adinardi, ejanzer, tony-dinh

Required Reviewers:

Approved By: kevinbarabash

Checks: ❌ codecov/project, ✅ lint_and_test (ubuntu-latest, 12), ✅ gerald, ✅ gerald, ✅ lint_and_test (ubuntu-latest, 12), ✅ gerald, ✅ lint_and_test (ubuntu-latest, 12), ✅ gerald, ✅ lint_and_test (ubuntu-latest, 12), ✅ gerald, ⏭  dependabot, ✅ chromatic (ubuntu-latest, 12)

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/1280