[fei4269.1.useServerEffect] Refactor WB Data to be less opinionated
This refactors our `IRequestHandler`-based architecture to remove that concept and instead have a more open, simpler API.
This replaces the "do everything and use state" `useData` hook with a "do just what we need" `useServerEffect` hook.

The client-side behavior of rehydrating and what have you is then part of the `Data` component. This makes a more generic hook for doing async server-based effects, while allowing for hydration cache support, without dictating how the result is ultimately used on the client-side. This opens up different caching options beyond just state.

We also add two additional props to `Data`; one to control whether a request is made after successful hydration, and the other to control whether the previous request's data is shown if a new request is made.