Add support for custom search filter to SingleSelect
When working on adding international school support to our School Picker that's part of the teacher dashboard, we found that some regions in international countries were named with accents. From testing, we figured out that the default search functionality of the SingleSelect dropdown component doesn't normalize the accents in these labels, and as a result, we're unable to allow users to search for "São Paulo" with the search string "Sao", where the "a" isn't accented.

Initially, the proposed solution was that we shouldn't need to match on the base letter only as our users can self-input the accents when appropriate, especially since they would be familiar with the accents as those would be their home regions. Unfortunately, we found that because of the way that the dropdown is rendered when we have more then 500 items, the focus gets shifted as the dropdown component is filtered, and the re-focusing of the element prevents our users for inputting the key-combination required for setting the accent (option-n on mac).

I've attached a screen recording displaying this behavior, but I figured if this component had a general-purpose solution to override the filtering logic, I could normalize the search text and the labels for each `OptionItem` so we can enable base-letter matching. This PR aims to accomplish this by adding a new `searchFilter` property to the component.

Questions for Reviewers:
- Looking at the confluence docs [1], I see that we're supposed to run `yarn changeset` anytime the public API is altered for a package, but this generated a markdown file in the `.changeset` directory. Should I be committing that file as part of this change?

[1] https://khanacademy.atlassian.net/wiki/spaces/WB/pages/100827261/Developing+Web+Wonder+Blocks

Issue: "none"

Test plan:

```shell
yarn test
```