Remove Styleguidist deprecated docs
Now that we have migrated most of the Web WB package docs to Storybook,
we can start deprecating all the Styleguidist docs that are now fully
migrated to Storybook.

This is based on the feedback provided here:

https://khanacademy.slack.com/archives/C8Z9DSKC7/p1663694718583589

Test Plan

Verify that the Styleguidist (Netlify) site is updated and doesn't throw
any error.