Revert "Version Packages (#1610)"
This reverts commit 0f546eba220c0f004c8b85e17c2561889ec5cd79.

The release failed b/c there was a changeset file with an invalid header.  I think it was the empty changeset I added in a recent PR that was updating tests and thus didn’t need to bump anything.  There were no errors for the checks in the PR, but there was an error in the Release workflow: https://github.com/Khan/wonder-blocks/actions/runs/3332349504/jobs/5513122456.