WB-1145.2: Allow more flexibility in Toolbar component
We noticed that the current `Toolbar` layout doesn't allow for any of
the content nodes to use all the available space.

This PR modifies the `Toolbar` layout to allow doing this.

See Slack thread: https://khanacademy.slack.com/archives/C8Z9DSKC7/p1667842562518939

While I was at that, I also did a bit of refactoring to reduce the
number of elements rendered and also updated the styles to use WB tokens
(e.g. Spacing and Color).

Test Plan:

Navigate to the `Toolbar > Responsive` story and resize the window to
verify that the content flows as expected.

1. In `small` screens, the main button uses all the available space and
   the secondary button is hidden.
2. In `large` screens, the main button is inlined, and another button is
   displayed as well.