Dropdown: Add `autoFocus` and `enableTypeAhead` props to `SingleSelect`.
Some users have reported some keyboard issues with the `TimePicker`
component created in webapp. This component is used to be able to either
select a time slot or type in a specific time.

This PR adds a new prop called `autoFocus` to the `SingleSelect`
component so we can prevent the input focus from being stolen by the
listbox.

Also, this PR adds a new prop called `enableTypeAhead` to the
`SingleSelect` component to avoid the listbox from being automatically
focused when the user types in the input.

Test Plan:

Navigate to the `AutoFocus` story and verify that the input works as
expected. This means that if the user focuses on the input and/or if the
user stars typing, the listbox should not be focused.