WB-1170: Add Storybook interaction testing to snapshot tooltip stories correctly
NOTE: This PR is based on #1599.

Introduces Storybook interaction testing to be able to create stable
tooltip snapshots in Chromatic.

This feature will help us to introduce more interaction tests that will
be used in Chromatic where we need to use focus/hover behaviors to
verify that some of the components are working correctly (e.g. Modals,
Popovers, Dropdowns, etc).

Test Plan:

Verify that the tooltip stories are correctly snapshotted in Chromatic.