Port Enzyme tests to RTL (`TextField`, `LabeledTextField`, `FieldHeading`)
Summary:

This is a port of the Enzyme tests for `TextField`, `LabeledTextField`, and `FieldHeading` to RTL.

I've done it as a preparatory step to making some layout changes in the
`FieldHeading` component. I figured out that it would be nice to have
RTL tests in place before making those changes.

Test Plan:

Make sure that all unit tests pass.