Add overview page to storybook
Adds an overview page to storybook. This page is intended to be the
first page that a user sees when they open storybook. It contains a
brief description of the project and some relevant links.

Also updates the README to point to the storybook page, update some
outdated links and use the GH actions badge.

Test plan:

Verify that the `?path=/story/overview--page` page makes sense.