WB-1081: TextField - Replace all "internal" Class Names to intended Names
We had to use "internal" class names for the components used in
`TextField` and `LabeledTextField` because Styleguidist was not able to
resolve the names properly (this was an issue with `Styleguidist` +
`forwardRef`).

Now that Styleguidist is gone, we can replace the "internal" class names
with the intended ones, as this no longer causes any issues with
Storybook.

NOTE: This is a super small change that will help us with better
readability when using React DevTools.

Test Plan:

Verify that all the checks pass (specially unit tests) and that the
`TextField` and `LabeledTextField` stories work fine.