WB-1190: Cleanup WB dependencies between packages
We noticed a while that we have an incorrect dependency in wb-dropdown.
I checked and we were not using wb-button internally in the code (only
for stories/documentation), so that might be bumping the dropdown
version when button changed unnecessarily.

This PR changes the other packages that are out-of-date to only include
the WB deps that are used in the code (excluding tests, stories, etc) so
the releases are a bit more contained.

In VS Code, I checked each `package.json` individually and used the
search panel to look for the other packages. I then checked the
`package.json` of the package I was looking for and compared any missing
WB packages that were not included in package.json, or removed the ones
that were not being imported in the code.

NOTE: I filtered out the contents of the `__tests__` and `__docs__`
folders for this search.