WB-1190: Cleanup WB dependencies between packages (#1735)
## Summary:
We noticed a while that we have an incorrect dependency in wb-dropdown.
I checked and we were not using wb-button internally in the code (only
for stories/documentation), so that might be bumping the dropdown
version when button changed unnecessarily.

This PR changes the other packages that are out-of-date to only include
the WB deps that are used in the code (excluding tests, stories, etc) so
the releases are a bit more contained.

In VS Code, I checked each `package.json` individually and used the
search panel to look for the other packages. I then checked the
`package.json` of the package I was looking for and compared any missing
WB packages that were not included in package.json, or removed the ones
that were not being imported in the code.

NOTE: I filtered out the contents of the `__tests__` and `__docs__`
folders for this search.

Issue: WB-1190

## Test plan:

Verify that the "Check builds" workflow job passes.

Author: jandrade

Reviewers: somewhatabstract

Required Reviewers:

Approved By: somewhatabstract

Checks: ❌ codecov/project, ✅ Test (ubuntu-latest, 16.x, 2/2), ✅ Test (ubuntu-latest, 16.x, 1/2), ✅ Lint (ubuntu-latest, 16.x), ✅ Check build sizes (ubuntu-latest, 16.x), ⏭  dependabot, ✅ Prime node_modules cache for primary configuration (ubuntu-latest, 16.x), ✅ gerald, ✅ Chromatic - Build on review PR (push) / chromatic (ubuntu-latest, 16.x), ⏭  Chromatic - Skip on dependabot PRs (push)

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/1735