Export the CJS build as the main entry point (#1739)
## Summary:
We recently removed Webpack from the build process, but we still need to
export the CJS build as the main entry point.

This is because the CJS build is the only one that works for some
particular cases such as when we use the `flow-enums-runtime` package.

In this Flow enums case, this happens because FlowJS needs to convert
the enums to code that can be executed at runtime. This is done by the
`babel-plugin-transform-flow-enums` plugin, which is configured to
transform enums from:

```
enum Foo {
  Bar,
  Baz,
};
```

to:

```
const Foo = require("flow-enums-runtime").Mirrored(["Bar", "Baz"]);
```

The problem with this is that the `flow-enums-runtime` package expects
us to import the CJS build, so it can import the module via `require`.

I attempted to fix this by changing the
`babel-plugin-transform-flow-enums` plugin to use a custom runtime, but
the problem is that the plugin is not fully configurable, so we cannot
change the output to something that allow us to staticly import the ESM
build.

e.g. (this won't work)
```
import { Mirrored } from "flow-enums-runtime";
const Foo = Mirrored(["Bar", "Baz"]);
```
The `getRuntime` is only capable of transforming the enum in a single
line.

This commit changes the `rollup.config.js` file to export the CJS build
as the main entry point, so we don't break the `flow-enums-runtime`
transform (that is used by Storybook in webapp).

Issue: XXX-XXXX

## Test plan:

Author: jandrade

Reviewers: jeresig

Required Reviewers:

Approved By: jeresig

Checks: ✅ codecov/project, ✅ Check build sizes (ubuntu-latest, 16.x), ✅ Test (ubuntu-latest, 16.x, 2/2), ✅ Test (ubuntu-latest, 16.x, 1/2), ✅ Lint (ubuntu-latest, 16.x), ⏭  Chromatic - Build on Release PR (PR opened), ✅ Prime node_modules cache for primary configuration (ubuntu-latest, 16.x), ⏭  dependabot, ✅ gerald, ✅ Chromatic - Build on review PR (push) / chromatic (ubuntu-latest, 16.x), ⏭  Chromatic - Skip on dependabot PRs (push)

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/1739