Fix the return type of useLatestRef
React defines the RefObject type as:

```
interface RefObject<T> {
  readonly current: T | null;
}
```

Previously, the `useLatestRef` hook was typed to return `RefObject<T>`,
but that wasn't really correct. The `current` value returned by
`useLatestRef` cannot be null if `T` is not nullable.

This commit changes the return type to `{readonly current: T}`, which is the
type returned by `React.useRef`.