WB-1590.3: Add @phosphor-icons/core support to Icon [prototype]
This PR adds support for the `@phosphor-icons/core` package to Icon.
This is a prototype PR, and it is not ready for review. It is intended
to be a starting point for discussion around how we want to support
Phosphor icons in WB Icon.

This is a follow up to #2014 which added support to
@phosphor-icons/react, and #2033, which added support to
@phosphor-icons/web.

Adds a new `PhosphorIcon` component that wraps an `img` element that
loads the SVG file imported from `@phosphor-icons/core`.