Fix filters for changeset workflow step
By default, filters are inclusive disjunctions (OR). This makes the `packages/` filter useless, since the glob filter will match anything that does not include `__tests__` and `dist` in the path.

This field that I added to the `filter-files` action will ensure that file pathnames match EVERY provided filter. This prevents accidental inclusions of packages outside of `packages/`.