Export `ModalContent` building block from modal package
There is a desire to export the `ModalContent` component to give consumers more control when constructing Modals using the building blocks. Currently, we always wrap the `ModalPanel`'s `content` prop in a `ModelContent` component wif it isn't already wrapped in one (which is typically the case); the `ModelContent` component has pre-defined styles and there ais currently no way to customize it, so we export the `ModelContent` component to allow for this as part of the Modal building blocks.

Issue: "none"

Test plan:

None (there is no functionality change)