WB-1607: Allow displaying the listbox based on the viewport size (#2136)
## Summary:

This PR changes the logic to display the listbox based on the viewport size.
Before this, we were calculating a fixed height for the listbox based on the
number of items in the listbox. This was causing the listbox to be displayed
partially and the scrollbar was not visible by default.

I'm changing that logic to display the listbox based on the viewport size. This
is done using a PopperJS modifier so we can detect the viewport size and
calculate the height of the listbox accordingly. By implementing this change, we
no longer display a fixed height and instead we display as many items as we can.

Issue: WB-1607

## Test plan:

Navigate to any of the `SingleSelect/MultiSelect` stories and verify that the
listbox is displayed based on the viewport size (resize the screen to see how
the listbox changes its height).


https://github.com/Khan/wonder-blocks/assets/843075/fee493a6-619b-4a16-b5bf-f59a40a6c467

Author: jandrade

Reviewers: nishasy, jandrade

Required Reviewers:

Approved By: nishasy

Checks: ✅ Chromatic - Get results on regular PRs (ubuntu-latest, 16.x), ❌ codecov/project, ✅ Test (ubuntu-latest, 16.x, 2/2), ✅ Check build sizes (ubuntu-latest, 16.x), ✅ Test (ubuntu-latest, 16.x, 1/2), ✅ Lint (ubuntu-latest, 16.x), ✅ Chromatic - Build on regular PRs / chromatic (ubuntu-latest, 16.x), ⏭  Chromatic - Skip on Release PR (changesets), ✅ Publish npm snapshot (ubuntu-latest, 16.x), ✅ Check for .changeset entries for all changed files (ubuntu-latest, 16.x), ✅ Prime node_modules cache for primary configuration (ubuntu-latest, 16.x), ✅ gerald, ⏭  dependabot

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2136