Add style and dropdownStyle props to BirthdayPicker
To allow the BirthdayPicker to be rendered in a wider variety of locations, such as narrow modals or mobile views, two new props have been added. "style" which allows the caller to add styles to the root element of the BirthdayPicker, and "dropdownStyle" which applies styles to the individual dropdowns. As opposed to a more narrow approach of specifying a "vertical" prop, this broad approach relieves the component of having to know when to render horizontally or vertically. This logic is placed in the callers hands, which can easily be determined via media query.

Issue: None

Test plan:
- View http://localhost:6061/?path=/story/packages-birthdaypicker--birthday-picker-vertical
- ensure component displays in a top to bottom manner
- View http://localhost:6061/?path=/story/packages-birthdaypicker--birthday-picker-vertical-with-error
- ensure component displays in the same manner, with an error message
- reran unit tests, ensure they pass