[fei6062.releaseprotections.2] Fail snapshot runs if a release is happening (#2414)
## Summary:
This brings over workflow changes from Perseus that are intended to prevent trying to publish snapshots if a release is in progress. This is the first layer of protection against this, we also have a script solution that works both in automation and locally for devs that will kick in for cases where this workflow change doesn't catch it.

This change looks for in-progress releases (`release.yml` runs for "Version Packages" changes) and fails the snapshot job if it finds one. This then adds a comment to the PR to indicate why the snapshot failed.

Issue: FEI-6062

## Test plan:
This was already tested for Perseus and testing requires orchestrating a release job at the same time as the PR trying to do snapshots. I don't think we need to do that for this PR, but we can if folks prefer.

Author: somewhatabstract

Reviewers: jandrade

Required Reviewers:

Approved By: jandrade

Checks: ⌛ Publish npm snapshot (ubuntu-latest, 20.x), ⌛ Check for .changeset entries for all changed files (ubuntu-latest, 20.x), ⌛ Prime node_modules cache for primary configuration (ubuntu-latest, 20.x), ⌛ Chromatic - Build on regular PRs / chromatic (ubuntu-latest, 20.x), ⏭️  Chromatic - Skip on Release PR (changesets), ⌛ gerald, ⏭️  dependabot

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2414