Added useLastDayOfMonth prop to BirthdayPicker
BirthdayPicker currently use the first day of the month for the date if `monthYearOnly` prop is set to `true`. But for some use-cases developers may want to use the last day of the month so that a user who is still 17, for example, is not considered 18, if their birthday is anything after the 1st of the month.

The new prop `useLastDayOfMonth` works in conjunction with `monthYearOnly`. If both are set to true, the date passed to `onChange` will be the lastday of the month.