[🔥AUDIT🔥] Revert recent BirthdayPicker picker changes (#2490)
🖍 _This is an audit!_ 🖍

## Summary:

Reverts this PR: #2486

The reason for reverting is to keep the current behavior of the BirthdayPicker
component, which is to default to the first day of the month when no day is
selected. This is to release other WB packages separately, and let this change
be released separately.

Issue: "none"

## Test plan:

N/A

Author: jandrade

Auditors: #wonder-blocks, dmcalpin

Required Reviewers:

Approved By:

Checks: ✅ 12 checks were successful, ⏭️  2 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2490