[dmcalpin/birthday-picker-update] Update monthYearOnly to use last day of month
BirthdayPicker currently use the first day of the month for the date if monthYearOnly prop is set to true. But for most use-cases developers want to use the last day of the month so that a user who is still 17, for example, is not considered 18, if their birthday is anything after the 1st of the month.

This update modifies the behavior of the component to pass the last day of the month to onChange if monthYearOnly is set to true.

Issue: PG-5959

Test plan:
- Modified unit tests to prove last day of month is returned when the prop is set