[WB-1902.2] IconButton: Update IconButton focus state to use box-shadow + outline for better a11y (#2529)
## Summary:

In #2524, we updated the IconButton focus state to use a box-shadow instead of
an outline. This change was made to improve accessibility and provide a more
modern look to the component.

Turns out that the box-shadow is not 100% accessible for all users, as it
doesn't work well on Windows High Contrast mode. This PR modifies the focus
style to use the `focus` style from `wonder-blocks-styles`, which is a
combination of a box-shadow and an outline. This should provide a more
consistent experience across different platforms and improve accessibility.

Issue: WB-1902

## Test plan:

Navigate to the IconButton component in Storybook and verify that the focus
state looks correct.

`/?path=/story/packages-iconbutton-all-variants--focus&globals=viewport:desktop;pseudo.focusVisible:!true`

Also verify that the close button in the Modal component has the correct focus
state.

`/?path=/story/packages-modal-onepanedialog--with-launcher`

Author: jandrade

Reviewers: beaesguerra

Required Reviewers:

Approved By: beaesguerra

Checks: ✅ 22 checks were successful, ⏭️  3 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2529