NavigationTabs: Add `animated` prop to enable transition animation (#2521)
## Summary:

If the `animated` prop is set to `true`, the current tab underline will transition to the tab item that is selected

We achieve this animation by moving the current indicator from NavigationTabItem to NavigationTabs. This is so that the indicator can move fluidly to the other tab items. In order to know what size and position to use for the underline, we dynamically determine the width and position of the current tab. We also use a ResizeObserver to update the width and position in case the sizing of the tabs changes. This also addresses the initial layout shift that occurs when the Aphrodite styles are first initialized.

Issue: WB-1906

## Test plan:
`?path=/story/packages-tabs-navigationtabs-navigationtabs--interactive`
- Confirm setting `animated` to `false` disables the transition and the current tab is still shown
- Confirm setting `animated` to `true` enables the current tab item transition

`?path=/story/packages-tabs-navigationtabs-navigationtabs-navigationtabs-all-variants--default`

Manually confirm the styling for NavigationTabs states (these were generating flaky snapshots in Chromatic and have been temporarily disabled. I reached out to the Chromatic team for support and created WB-1917 as a follow up!)

- hover
- focus
- hover + focus
- active
- zoom

Author: beaesguerra

Reviewers: beaesguerra, marcysutton, jandrade

Required Reviewers:

Approved By: jandrade

Checks: ✅ 12 checks were successful, ⏭️  2 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2521