[WB-1852.1] IconButton: Remove `light` variant (#2530)
## Summary:

This PR removes the `light` variant from the IconButton component and replaces
it with a new `actionStyles.inverse` style. This change is part of a larger
effort to standardize the styling of interactive elements across WB. The `light`
variant was previously used in the `Popover` and `Moda`l components. The new
`actionStyles.inverse` style provides a way to apply almost the same styling as
the `light` variant, but allow us to avoid having to maintain a separate `light`
variant in the IconButton component (which won't be supported in Thunder
Blocks and new styles).

Key changes:

- Styles: Added a new `actionStyles.inverse` style to apply to interactive
(action) components.
- IconButton: Removed the `light` variant and replaced it with
`actionStyles.inverse`.
- Modal and Popover: Updated dismiss button (CloseButton) to use
`actionStyles.inverse` instead of the now deprecated `IconButton.light` variant.

Issue: WB-1852

## Test plan:

### IconButton

Navigate to the `IconButton` docs and verify that all the dark stories are no
longer present. Also verify that the "All variant" snapshots remove the `light`
variant.

- /?path=/docs/packages-iconbutton--docs
- /?path=/story/packages-iconbutton-all-variants--default


### Popover and Modal

Verify that the `Popover` and `Modal` components are still working as expected
and that the dismiss button (CloseButton) is using the new
`actionStyles.inverse` style.

- Popover: /?path=/docs/packages-popover-popovercontent--docs#with%20illustration
- Modal: /?path=/story/packages-modal-building-blocks-modalheader--dark&globals=viewport:desktop;

### Styles

Verify that the new `actionStyles.inverse` style is working as expected in the
`ActionStyles` docs.

- /?path=/docs/packages-styles-action-styles--docs

Author: jandrade

Reviewers: beaesguerra, jandrade

Required Reviewers:

Approved By: beaesguerra

Checks: ✅ 12 checks were successful, ⏭️  2 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2530